# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "AccessOperationsDto": {
        "additionalProperties": false,
        "properties": {
          "MANAGE_WEBHOOKS": {
            "default": false,
            "type": "boolean"
          },
          "READ": {
            "default": true,
            "type": "boolean"
          },
          "SEND_MESSAGES": {
            "default": false,
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "CreateServerDto": {
        "additionalProperties": false,
        "properties": {
          "defaultPermissionPolicy": {
            "default": "deny_all",
            "enum": [
              "deny_all",
              "allow_all",
              "custom"
            ],
            "example": "deny_all",
            "type": "string"
          },
          "disabledReason": {
            "example": "maintenance",
            "nullable": true,
            "type": "string"
          },
          "guildId": {
            "description": "Discord guild ID",
            "example": "123456789012345678",
            "type": "string"
          },
          "icon": {
            "example": "https://cdn.discordapp.com/icons/.../icon.png",
            "nullable": true,
            "type": "string"
          },
          "isActive": {
            "default": true,
            "example": true,
            "type": "boolean"
          },
          "isMain": {
            "default": false,
            "example": false,
            "type": "boolean"
          },
          "name": {
            "example": "Test Guild",
            "type": "string"
          },
          "syncFrequencyHours": {
            "description": "Sync frequency in hours",
            "example": 2,
            "minimum": 1,
            "type": "number"
          },
          "type": {
            "default": "other",
            "description": "Server category type",
            "enum": [
              "main",
              "competition",
              "event",
              "other"
            ],
            "example": "other",
            "type": "string"
          }
        },
        "required": [
          "guildId"
        ],
        "type": "object"
      },
...
      "DisableServerDto": {
        "additionalProperties": false,
        "properties": {
          "disabledReason": {
            "example": "maintenance",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "SetProjectServerAccessDto": {
        "additionalProperties": false,
        "properties": {
          "operations": {
            "$ref": "#/components/schemas/AccessOperationsDto"
          },
          "scopes": {
            "description": "Scopes granted to this project for this server. If omitted, defaults to all available scopes.",
            "example": [
              "read_members",
              "check_permissions"
            ],
            "items": {
              "enum": [
                "read_members",
                "check_permissions"
              ],
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "UpdateServerDto": {
        "additionalProperties": false,
        "properties": {
          "defaultPermissionPolicy": {
            "default": "deny_all",
            "enum": [
              "deny_all",
              "allow_all",
              "custom"
            ],
            "example": "deny_all",
            "type": "string"
          },
          "icon": {
            "example": "https://cdn.discordapp.com/icons/.../icon.png",
            "nullable": true,
            "type": "string"
          },
          "isMain": {
            "default": false,
            "example": false,
            "type": "boolean"
          },
          "name": {
            "example": "Test Guild",
            "type": "string"
          },
          "syncFrequencyHours": {
            "description": "Sync frequency in hours",
            "example": 2,
            "minimum": 1,
            "type": "number"
          },
          "type": {
            "default": "other",
            "description": "Server category type",
            "enum": [
              "main",
              "competition",
              "event",
              "other"
            ],
            "example": "other",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "api-key": {
        "description": "Project API key (prefix.secret format)",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "session-token": {
        "bearerFormat": "Token",
        "description": "Enter your session token (from admin login or POST /auth/validate)",
        "in": "header",
        "name": "Authorization",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "MicroClub Discord Identity API.\n\n**MC Project endpoints** authenticate with `X-API-Key` header.\n\n**Admin endpoints** authenticate with `Authorization: Bearer <token>`.\n\nScopes and operations are granted per project–server pair by a system administrator.",
    "title": "MCDI API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/projects/{projectId}/servers/{serverId}": {
      "delete": {
        "operationId": "ProjectsController_revokeProjectServerAccess",
        "parameters": [
          {
            "description": "Project UUID",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Discord server ID",
            "in": "path",
            "name": "serverId",
            "required": true,
            "schema": {
              "example": "123456789012345678",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Access mapping revoked."
          },
          "400": {
            "description": "Invalid project or server ID format."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "System Admin access required."
          },
          "404": {
            "description": "Access mapping not found."
          }
        },
        "security": [
          {
            "session-token": []
          }
        ],
        "summary": "Revoke project access to a server",
        "tags": [
          "Projects"
        ]
      },
      "put": {
        "description": "Creates or updates the access mapping between a project and a server. You can set both the allowed operations and the scopes per server. If scopes are omitted, existing scopes are preserved (or all scopes are granted for new mappings).",
        "operationId": "ProjectsController_setProjectServerAccess",
        "parameters": [
          {
            "description": "Project UUID",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Discord server ID",
            "in": "path",
            "name": "serverId",
            "required": true,
            "schema": {
              "example": "123456789012345678",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetProjectServerAccessDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "createdAt": "2026-02-20T00:00:00.000Z",
                    "operations": {
                      "MANAGE_WEBHOOKS": false,
                      "READ": true,
                      "SEND_MESSAGES": false
                    },
                    "projectId": "123e4567-e89b-12d3-a456-426614174000",
                    "scopes": [
                      "read_members",
                      "check_permissions"
                    ],
                    "serverId": "123456789012345678",
                    "updatedAt": "2026-03-09T12:00:00.000Z"
                  }
                }
              }
            },
            "description": "Access mapping upserted."
          },
          "400": {
            "description": "Invalid project or server ID format."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "System Admin access required."
          },
          "404": {
            "description": "Project or server not found."
          }
        },
        "security": [
          {
            "session-token": []
          }
        ],
        "summary": "Grant or update project access to a server",
        "tags": [
          "Projects"
        ]
      }
    },
...
    "/api/servers": {
      "get": {
        "description": "Returns all registered servers with their current sync health data.",
        "operationId": "ServersController_list",
        "parameters": [
          {
            "description": "Filter by active/inactive status.",
            "in": "query",
            "name": "isActive",
            "required": false,
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "description": "Filter by main server status.",
            "in": "query",
            "name": "isMain",
            "required": false,
            "schema": {
              "example": false,
              "type": "boolean"
            }
          },
          {
            "description": "Filter by server type (e.g. official, partner, other).",
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "example": "official",
              "maxLength": 50,
              "type": "string"
            }
          },
          {
            "description": "Filter by server name (case-insensitive partial match).",
            "in": "query",
            "name": "name",
            "required": false,
            "schema": {
              "example": "MicroClub",
              "maxLength": 255,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Servers retrieved successfully."
          },
          "400": {
            "description": "Invalid parameter format."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "System admin access required."
          }
        },
        "security": [
          {
            "session-token": []
          }
        ],
        "summary": "List all servers",
        "tags": [
          "Servers"
        ]
      },
      "post": {
        "description": "Registers a new Discord guild with MCDI (upsert by guild ID). Once registered, the server can be granted to projects and included in syncs.",
        "operationId": "ServersController_register",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServerDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Server registered successfully."
          },
          "400": {
            "description": "Invalid request body."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "System admin access required."
          }
        },
        "security": [
          {
            "session-token": []
          }
        ],
        "summary": "Register a server",
        "tags": [
          "Servers"
        ]
      }
    },
    "/api/servers/{serverId}": {
      "delete": {
        "description": "Permanently removes a server and all associated data (members, roles, sync logs). This cannot be undone.",
        "operationId": "ServersController_remove",
        "parameters": [
          {
            "description": "Discord guild snowflake ID",
            "in": "path",
            "name": "serverId",
            "required": true,
            "schema": {
              "example": "123456789012345678",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Server deleted successfully."
          },
          "400": {
            "description": "Invalid server ID format."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "System admin access required."
          },
          "404": {
            "description": "Server not found."
          }
        },
        "security": [
          {
            "session-token": []
          }
        ],
        "summary": "Delete a server",
        "tags": [
          "Servers"
        ]
      },
      "get": {
        "description": "Returns a single server record by its Discord guild snowflake ID.",
        "operationId": "ServersController_findOne",
        "parameters": [
          {
            "description": "Discord guild snowflake ID",
            "in": "path",
            "name": "serverId",
            "required": true,
            "schema": {
              "example": "123456789012345678",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server retrieved successfully."
          },
          "400": {
            "description": "Invalid server ID format."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "System admin access required."
          },
          "404": {
            "description": "Server not found."
          }
        },
        "security": [
          {
            "session-token": []
          }
        ],
        "summary": "Get a server by ID",
        "tags": [
          "Servers"
        ]
      },
      "patch": {
        "description": "Updates the server name or other mutable settings.",
        "operationId": "ServersController_update",
        "parameters": [
          {
            "description": "Discord guild snowflake ID",
            "in": "path",
            "name": "serverId",
            "required": true,
            "schema": {
              "example": "123456789012345678",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServerDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Server updated successfully."
          },
          "400": {
            "description": "Invalid request body."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "System admin access required."
          },
          "404": {
            "description": "Server not found."
          }
        },
        "security": [
          {
            "session-token": []
          }
        ],
        "summary": "Update server settings",
        "tags": [
          "Servers"
        ]
      }
    },
...
    "/api/servers/{serverId}/disable": {
      "patch": {
        "description": "Marks the server as inactive. Disabled servers are excluded from syncs and all API permission/member checks return 403.",
        "operationId": "ServersController_disable",
        "parameters": [
          {
            "description": "Discord guild snowflake ID",
            "in": "path",
            "name": "serverId",
            "required": true,
            "schema": {
              "example": "123456789012345678",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisableServerDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Server disabled successfully."
          },
          "400": {
            "description": "Invalid server ID format."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "System admin access required."
          },
          "404": {
            "description": "Server not found."
          }
        },
        "security": [
          {
            "session-token": []
          }
        ],
        "summary": "Disable a server",
        "tags": [
          "Servers"
        ]
      }
    },
    "/api/servers/{serverId}/enable": {
      "patch": {
        "description": "Re-activates a previously disabled server.",
        "operationId": "ServersController_enable",
        "parameters": [
          {
            "description": "Discord guild snowflake ID",
            "in": "path",
            "name": "serverId",
            "required": true,
            "schema": {
              "example": "123456789012345678",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server enabled successfully."
          },
          "400": {
            "description": "Invalid server ID format."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "System admin access required."
          },
          "404": {
            "description": "Server not found."
          }
        },
        "security": [
          {
            "session-token": []
          }
        ],
        "summary": "Enable a server",
        "tags": [
          "Servers"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "API Server",
      "url": "http://localhost:3000"
    }
  ],
...
}